# Conflicts:
#	services/Vch/NTPTools/ITimeProvider.cs
#	services/Vch/NTPTools/TimeProvider.cs
#	services/Vch/Vch.Checker/Program.cs
#	services/Vch/Vch.Core/Sorages/BaseMongoStorage.cs
#	services/Vch/Vch.Core/Sorages/MessageStorage.cs

+++ services/Vch/Vch.Core/Helpers/BinaryExtensoins.cs
@@ -19,7 +19,7 @@ namespace Vch.Core.Helpers
             }
         }
 
-        public static byte[] ToBytes(this double source)
+        public static byte[] ToBytes(this ulong source)
         {
             return BitConverter.GetBytes(source);
         }
